+++ app/src/main/java/ar/scacchipa/jsonplaceholderviewer/app/MainActivity.kt
@@ -1,33 +1,13 @@
-package ar.scacchipa.jsonplaceholderviewer
+package ar.scacchipa.jsonplaceholderviewer.app
 
-import android.app.Application
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
 import androidx.recyclerview.widget.LinearLayoutManager
 import ar.scacchipa.jsonplaceholderviewer.databinding.ActivityMainBinding
+import ar.scacchipa.jsonplaceholderviewer.ui.CommentAdapter
+import ar.scacchipa.jsonplaceholderviewer.ui.MiddleCommentViewModel
 import org.koin.android.ext.android.inject
-import org.koin.androidx.viewmodel.dsl.viewModel
-import org.koin.core.context.startKoin
-import org.koin.dsl.module
 
-val appModule = module {
-
-    factory { provideRetrofit() }
-    single <IJPHSourceData> { provideNetworkApi(get()) }
-    single <IJPHRepository> { JPHRepository( get() ) }
-    single <MiddleCommentsUserCase> { MiddleCommentsUserCase( get() ) }
-    viewModel { MiddleCommentViewModel( get() ) }
-}
-
-class MyApplication: Application() {
-    override fun onCreate() {
-        super.onCreate()
-        startKoin {
-            //androidContext(this@MyApplication)
-            modules(appModule)
-        }
-    }
-}
 class MainActivity : AppCompatActivity() {
 
     private val commentVM: MiddleCommentViewModel by inject()
@@ -48,6 +28,5 @@ class MainActivity : AppCompatActivity() {
                 binding.commentRecyclerView.adapter = CommentAdapter(it)
             }
         }
-
     }
 }
