+++ app/src/main/java/ar/scacchipa/jsonplaceholderviewer/app/MyApplication.kt
@@ -1,0 +1,14 @@
+package ar.scacchipa.jsonplaceholderviewer.app
+
+import android.app.Application
+import org.koin.core.context.startKoin
+
+class MyApplication: Application() {
+    override fun onCreate() {
+        super.onCreate()
+        startKoin {
+            //androidContext(this@MyApplication)
+            modules(appModule)
+        }
+    }
+}
