+++ app/src/main/java/ar/scacchipa/jsonplaceholderviewer/data/Comment.kt
@@ -1,4 +1,4 @@
-package ar.scacchipa.jsonplaceholderviewer
+package ar.scacchipa.jsonplaceholderviewer.data
 
 data class Comment(
     val postId: Int,
