+++ app/src/main/java/ar/scacchipa/jsonplaceholderviewer/domain/MiddleCommentsUserCase.kt
@@ -1,10 +1,12 @@
-package ar.scacchipa.jsonplaceholderviewer
+package ar.scacchipa.jsonplaceholderviewer.domain
 
+import ar.scacchipa.jsonplaceholderviewer.data.Comment
+import ar.scacchipa.jsonplaceholderviewer.data.ICommentRepository
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.withContext
 
 class MiddleCommentsUserCase(
-    private val jphRepository: IJPHRepository// = JPHRepository()
+    private val jphRepository: ICommentRepository// = JPHRepository()
 ) {
     suspend fun getMiddleCommentList(): List<Comment> =
         withContext(Dispatchers.IO) {
