+++ app/src/main/java/ar/scacchipa/jsonplaceholderviewer/ui/MiddleCommentViewModel.kt
@@ -1,8 +1,10 @@
-package ar.scacchipa.jsonplaceholderviewer
+package ar.scacchipa.jsonplaceholderviewer.ui
 
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
+import ar.scacchipa.jsonplaceholderviewer.data.Comment
+import ar.scacchipa.jsonplaceholderviewer.domain.MiddleCommentsUserCase
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
